Move dynamic_ir to ts_backend
